packages/ingestkit-email/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "ingestkit-email"
version = "0.1.0"
description = "Email (EML/MSG) document converter for RAG pipelines"
requires-python = ">=3.10"
dependencies = [
    "ingestkit-core>=0.1.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
# MSG format support
msg = ["extract-msg>=0.48.0"]

# Vector store backends
qdrant = ["qdrant-client>=1.7"]

# LLM / Embedding backends
ollama = ["httpx>=0.27"]

# Convenience bundles
all = ["ingestkit-email[msg,qdrant,ollama]"]

# Development
dev = [
    "extract-msg>=0.48.0",
    "pytest>=7.0",
    "pytest-cov",
    "pyyaml>=6.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: Unit tests (no external services)",
    "integration: Integration tests (require external services)",
]
